# GTEST/UCT: Do not check gdr_copy perf in test_uct_loopback